friends and mentors are now sorted into the list on the main screen based on a new parameter called friendPoints
Former-commit-id: f36e89c039a006ef2ea0bd65fef850067d119eee

+++ Assets/Scripts/UI/HomeworkIndicator.cs
@@ -42,6 +42,6 @@ public class HomeworkIndicator : MonoBehaviour {
             indicatorText.text = "HELP!";
         }
 
-        indicatorText.color = Color.Lerp(noHomeworkColor, maxHomeworkColor, (float) (level/maxHomework));
+        indicatorText.color = Color.Lerp(noHomeworkColor, maxHomeworkColor, (float) ((float) level/(float) maxHomework));
     }
 }
